+++ packages/flutter_form_builder/lib/src/form_builder.dart
@@ -60,6 +60,9 @@ class FormBuilder extends StatefulWidget {
   /// and their enabled state will be ignored.
   final bool enabled;
 
+  /// Whether the form should auto focus on the first field that fails validation.
+  final bool autoFocusOnValidationFailure;
+
   /// Creates a container for form fields.
   ///
   /// The [child] argument must not be null.
@@ -72,6 +75,7 @@ class FormBuilder extends StatefulWidget {
     this.initialValue = const <String, dynamic>{},
     this.skipDisabled = false,
     this.enabled = true,
+    this.autoFocusOnValidationFailure = false,
   }) : super(key: key);
 
   static FormBuilderState? of(BuildContext context) =>
@@ -214,7 +218,7 @@ class FormBuilderState extends State<FormBuilder> {
 
   bool validate() {
     final hasError = !_formKey.currentState!.validate();
-    if (hasError) {
+    if (hasError && widget.autoFocusOnValidationFailure) {
       final wrongFields =
           fields.values.where((element) => element.hasError).toList();
       wrongFields.first.requestFocus();